perf(bundle): move yjs-dependent helpers out of utils.ts into yjs-utils (W2)
Relocate extractMarkdownFromYjs and extractPreviewTextFromYjs (with their
yjs import and getDocumentUpdates import) into src/lib/yjs-utils.ts so that
the ~43 files importing only cn() and light helpers from @/lib/utils no
longer transitively pull the entire yjs library. Three importers rewired:
ShareDialog.tsx, ShareProvider.ts, SyncService.ts (split import line).
